Major updates across the board. - replaced the Model project (edmx) with the Orvosi.Data project (code first) - too many changes to document. everything is going to be retested. - added ServiceRequestTemplates - Changed the dependsOn CSV column to a related table (Dependent tables)

+++ Databases/Orvosi/dbo/Tables/RoleCategory.sql
@@ -1,10 +1,12 @@
 ﻿CREATE TABLE [dbo].[RoleCategory] (
-    [Id]           SMALLINT       NOT NULL,
+    [Id]           TINYINT        NOT NULL,
     [Name]         NVARCHAR (128) NOT NULL,
     [ModifiedDate] DATETIME       CONSTRAINT [DF_RoleCategory_ModifiedDate] DEFAULT (getdate()) NOT NULL,
     [ModifiedUser] NVARCHAR (100) CONSTRAINT [DF_RoleCategory_ModifiedUser] DEFAULT (suser_name()) NOT NULL,
     CONSTRAINT [PK_RoleCategory] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
+
+
